+++ db_setup/db_create.sql
@@ -36,7 +36,7 @@ CREATE TABLE item (
 	reference TEXT UNIQUE NOT NULL,
 	packaging_id INT,
 	category_id INT NOT NULL,
-	last_edit DATE NOT NULL,
+	last_edit TIMESTAMP WITH TIME ZONE NOT NULL,
 	FOREIGN KEY(packaging_id) REFERENCES packaging(id),
 	FOREIGN KEY(category_id) REFERENCES category(id));