+++ src/api/routers/index.js
@@ -1,6 +1,15 @@
 import itemRouter from './item';
 import categoryRouter from './category';
 import propertyRouter from './property';
+import requestRouter from './request';
+import packagingRouter from './packaging';
 import userRouter from './user';
 
-export default [itemRouter, categoryRouter, propertyRouter, userRouter];
+export default [
+  itemRouter,
+  categoryRouter,
+  propertyRouter,
+  requestRouter,
+  packagingRouter,
+  userRouter
+];