+++ src/api/routers/item.js
@@ -42,7 +42,8 @@ itemRouter.post('/add_new_item', async (req, res) => {
 });
 
 itemRouter.get('/item_characteristics/:id', async (req, res) => {
-  const item_description_query = `SELECT item.description, convert_from(item.image, 'UTF-8') as image, item.total_stock, item.free_stock, item.last_price, item.location, item.user_comments, item.details, item.manufacturer, item.reference, item.packaging
+  const item_description_query = `SELECT item.description, convert_from(item.image, 'UTF-8') as image, item.total_stock, item.free_stock,
+  item.last_price, item.location, item.user_comments, item.details, item.manufacturer, item.reference, to_char(item.last_edit, 'DD Mon YYYY HH24hMIm') as last_edit
     FROM item
     WHERE item.id = ${req.params.id}`;
 
@@ -57,14 +58,12 @@ itemRouter.get('/item_characteristics/:id', async (req, res) => {
 
 itemRouter.get('/item_properties/:id', async (req, res) => {
   const in_depth_item_ppt_list_query = `
-    SELECT item_property.property_id, item_property.value, property.unit, property.name
+    SELECT item_property.property_id, item_property.value, property.unit, property.name, property.number
     FROM item_property, property
-    WHERE property.id = item_property.property_id and item_property.item_id = ${
-      req.params.id
-    }`;
+    WHERE property.id = item_property.property_id and item_property.item_id = $1`;
 
   try {
-    const data = await db.any(in_depth_item_ppt_list_query);
+    const data = await db.any(in_depth_item_ppt_list_query, req.params.id);
     res.send(data);
   } catch (e) {
     console.log('Error retrieving item properties! ', e);
@@ -88,9 +87,25 @@ itemRouter.get('/item_category/:id', async (req, res) => {
   }
 });
 
+itemRouter.get('/item_packaging/:id', async (req, res) => {
+  const in_depth_item_packaging_query = `
+    SELECT packaging.id, packaging.name
+    FROM packaging, item
+    WHERE packaging.id = item.packaging_id
+      AND item.id = ${req.params.id}`;
+
+  try {
+    const data = await db.one(in_depth_item_packaging_query);
+    res.send(data);
+  } catch (e) {
+    console.log('Error retrieving item packaging! ', e);
+    res.send('Failed to retrieve item packaging!');
+  }
+});
+
 itemRouter.post('/item_edit', async (req, res) => {
   let newItem = req.body;
-  console.log('itemedit', newItem);
+
   let item_update_query = `UPDATE item
     SET description = '${newItem.description}',
         image = '${newItem.image}',
@@ -102,8 +117,9 @@ itemRouter.post('/item_edit', async (req, res) => {
         details = '${newItem.details}',
         manufacturer = '${newItem.manufacturer}',
         reference = '${newItem.reference}',
-        packaging = '${newItem.packaging}',
-        category_id = ${newItem.category.itemCategory.id}
+        packaging_id = ${newItem.packaging.itemPackaging.id},
+        category_id = ${newItem.category.itemCategory.id},
+        last_edit = NOW()
     WHERE id = ${newItem.id}; \n`;
 
   for (let i = 0; i < newItem.properties.length; i++) {
@@ -113,7 +129,6 @@ itemRouter.post('/item_edit', async (req, res) => {
       VALUES ('${property.value}', ${newItem.id}, ${property.property_id}); \n`;
 
     item_update_query = item_update_query.concat(property_update_query);
-    console.log('query', property_update_query);
   }
 
   try {