+++ CHANGELOG.md
@@ -1,7 +1,21 @@
 ## 0.9.0 (13/10/2018)
 
 -   Full view of available components page
-    
+
 -   Single component detailed page
-    
+
 -   Editing the single component detailed page
+
+## (30/10/2018)
+
+## (13/11/2019)
+
+-   What was incremented during Sprint 3:
+
+-   Change the order of the columns
+-   Hide columns
+-   Select items to compare side by side
+-   Breadcrumbs to show the categories selected (if a selected category is a child of the previous, it will appear at right)
+-   Three types of login (professor, student and manager)
+
+-   We closed the issues: #25, #27, #35, #36, #37, #38, #40, #41.